refactor: 클라이언트 뷰 컴포넌트 import 경로 개선 및 정리
